Grouped interaction token types
