+++ package.json
@@ -18,7 +18,7 @@
     ],
     "exclude": [
       "./**/types.ts",
-      "./**/*Types.ts"
+      "./**/*.types.ts"
     ],
     "extension": [
       ".ts"
