+++ ts/interactionTokens/credentialOffer.ts
@@ -1,5 +1,5 @@
 import { classToPlain, plainToClass, Expose, Exclude } from 'class-transformer'
-import { ICredentialOfferCreationAttrs } from './credentialOfferTypes'
+import { ICredentialOfferCreationAttrs } from './interactionTokens.types'
 
 /* Class representing a credential offer. Encodable in JWT */
