--- ts/interactionTokens/credentialOfferTypes.ts
@@ -1,8 +1,0 @@
-export interface ICredentialOfferCreationAttrs {
-  challenge: string
-  callbackURL: string
-  instant: boolean
-  requestedInput: {
-    [key: string]: string | null
-  }
-}
