--- ts/interactionTokens/credentialResponseTypes.ts
@@ -1,6 +1,0 @@
-import { ISignedCredentialAttrs } from '../credentials/signedCredential/types'
-
-export interface ICredentialResponseAttrs {
-  callbackURL: string
-  suppliedCredentials: ISignedCredentialAttrs[]
-}
