--- ts/interactionTokens/credentialsReceiveTypes.ts
@@ -1,4 +1,0 @@
-import { ISignedCredentialAttrs } from '../credentials/signedCredential/types'
-export interface ICredentialsReceiveAttrs {
-  signedCredentials: ISignedCredentialAttrs[]
-}
