+++ ts/interactionTokens/interactionTokens.types.ts
@@ -1,0 +1,59 @@
+import { ISignedCredentialAttrs } from "../credentials/signedCredential/types"
+
+/* Creation attributes. Same as JSON form */
+
+export interface ICredentialOfferAttrs {
+  challenge: string
+  callbackURL: string
+  instant: boolean
+  requestedInput: {
+    [key: string]: string | null
+  }
+}
+
+export interface ICredentialOfferCreationAttrs {
+  challenge: string
+  callbackURL: string
+  instant: boolean
+  requestedInput: {
+    [key: string]: string | null
+  }
+}
+
+export interface ICredentialRequestAttrs {
+  callbackURL: string
+  credentialRequirements: ICredentialRequest[]
+}
+
+export interface ICredentialResponseAttrs {
+  callbackURL: string
+  suppliedCredentials: ISignedCredentialAttrs[]
+}
+
+export interface ICredentialsReceiveAttrs {
+  signedCredentials: ISignedCredentialAttrs[]
+}
+
+export interface ICredentialRequest {
+  type: string[]
+  constraints: IConstraint[]
+}
+
+/* Related to constraint functions */
+
+export type Operator = '==' | '!=' | '<' | '>'
+
+export interface IConstraint {
+  [operator: string]: boolean[] | Array<{ var: string } | string | Comparable>
+}
+
+export type Comparable = number | Date
+type ConstraintFunc = (field: string, value: string) => IConstraint
+type ComparableConstraintFunc = (field: string, value: Comparable) => IConstraint
+
+export interface IExposedConstraintFunctions {
+  is: ConstraintFunc
+  not: ConstraintFunc
+  greater: ComparableConstraintFunc
+  smaller: ComparableConstraintFunc
+}
